+++ backend/entityservice/utils.py
@@ -55,6 +55,63 @@ def iterable_to_stream(iterable, buffer_size=io.DEFAULT_BUFFER_SIZE):
     return io.BufferedReader(IterStream(), buffer_size=buffer_size)
 
 
+def chain_streams(streams, buffer_size=io.DEFAULT_BUFFER_SIZE):
+    """
+    Chain an iterable of streams together into a single buffered stream.
+
+    Usage:
+        def generate_open_file_streams():
+            for file in filenames:
+                yield open(file, 'rb')
+
+        f = chain_streams(generate_open_file_streams())
+        f.read()
+
+    """
+
+    class ChainStream(io.RawIOBase):
+        def __init__(self):
+            self.leftover = b''
+            self.stream_iter = iter(streams)
+            try:
+                self.stream = next(self.stream_iter)
+            except StopIteration:
+                self.stream = None
+
+        def readable(self):
+            return True
+
+        def _read_next_chunk(self, max_length):
+            # Return 0 or more bytes from the current stream, first returning all
+            # leftover bytes. If the stream is closed returns b''
+            if self.leftover:
+                return self.leftover
+            elif self.stream is not None:
+                return self.stream.read(max_length)
+            else:
+                return b''
+
+        def readinto(self, b):
+            buffer_length = len(b)
+            chunk = self._read_next_chunk(buffer_length)
+            while len(chunk) == 0:
+                # move to next stream
+                if self.stream is not None:
+                    self.stream.close()
+                try:
+                    self.stream = next(self.stream_iter)
+                    chunk = self._read_next_chunk(buffer_length)
+                except StopIteration:
+                    # No more streams to chain together
+                    self.stream = None
+                    return 0  # indicate EOF
+            output, self.leftover = chunk[:buffer_length], chunk[buffer_length:]
+            b[:len(output)] = output
+            return len(output)
+
+    return io.BufferedReader(ChainStream(), buffer_size=buffer_size)
+
+
 def safe_fail_request(status_code, message, **kwargs):
     """
     generates an error message in the right format.
@@ -119,3 +176,11 @@ def clks_uploaded_to_project(project_id):
     number_parties = get_project_column(conn, project_id, 'parties')
     logging.info("{}/{} parties have contributed clks".format(parties_contributed, number_parties))
     return parties_contributed == number_parties
+
+
+def similarity_matrix_from_csv_bytes(data):
+    rows = data.decode().splitlines()
+    for row in rows:
+        index_1, index_2, score = row.split(',')
+        # Note we rewrite in a different order because we love making work for ourselves
+        yield (int(index_1), float(score), int(index_2))